Refactor logging view into controller and view for better testability. Refactor table into view and data classes for better testability.
