+++ packages/devtools/test/goldens/inspector_controller_details_tree_scaffold_with_styles.txt
@@ -1,29 +1,29 @@
-▼[S]<grayed> </grayed><bold>Scaffold</bold> <-- selected
-└─▼[S]<grayed> </grayed>_ScaffoldScope
-    ▼[P]<grayed> </grayed>PrimaryScrollController
-    └─▼[M]<grayed> </grayed>Material
-      └─▼[/icons/inspector/resume.png]<grayed> </grayed>AnimatedPhysicalModel
-        └─▼[P]<grayed> </grayed>PhysicalModel
-          └─▼[N]<grayed> </grayed>NotificationListener <grayed><LayoutChangedNotification></grayed>
-              ▼[I]<grayed> </grayed>_InkFeatures <grayed>[GlobalKey#00000 ink renderer]</grayed>
-              └─▼[/icons/inspector/resume.png]<grayed> </grayed>AnimatedDefaultTextStyle
-                └─▼[/icons/inspector/textArea.png]<grayed> </grayed>DefaultTextStyle
-                  └─▼[/icons/inspector/resume.png]<grayed> </grayed>AnimatedBuilder
-                    └─▼[C]<grayed> </grayed>CustomMultiChildLayout
-                      ├───▼[L]<grayed> </grayed>LayoutId <grayed>[<_ScaffoldSlot.body>]</grayed>
-                      │   └─▼[/icons/inspector/atrule.png]<grayed> </grayed>MediaQuery
-                      │     └─▼[C]<grayed> </grayed><bold>Center</bold>
-                      │       └─▶[/icons/inspector/textArea.png]<grayed> </grayed><bold>Text</bold>
-                      ├───▼[L]<grayed> </grayed>LayoutId <grayed>[<_ScaffoldSlot.appBar>]</grayed>
-                      │   └─▼[/icons/inspector/atrule.png]<grayed> </grayed>MediaQuery
-                      │     └─▼[C]<grayed> </grayed>ConstrainedBox
-                      │       └─▼[F]<grayed> </grayed>FlexibleSpaceBarSettings
-                      │           ▶[A]<grayed> </grayed><bold>AppBar</bold>
-                      └─▼[L]<grayed> </grayed>LayoutId <grayed>[<_ScaffoldSlot.floatingActionButton>]</grayed>
-                        └─▼[/icons/inspector/atrule.png]<grayed> </grayed>MediaQuery
-                          └─▼[F]<grayed> </grayed>_FloatingActionButtonTransition
-                            └─▼[/icons/inspector/value.png]<grayed> </grayed>Stack
-                              └─▼[/icons/inspector/resume.png]<grayed> </grayed>ScaleTransition
-                                └─▼[/icons/inspector/colors.png]<grayed> </grayed>Transform
-                                  └─▼[/icons/inspector/resume.png]<grayed> </grayed>RotationTransition
-                                    └─▼[/icons/inspector/colors.png]<grayed> </grayed>Transform
+▼[S]<bold>Scaffold</bold> <-- selected
+└─▼[S]_ScaffoldScope
+    ▼[P]PrimaryScrollController
+    └─▼[M]Material
+      └─▼[/icons/inspector/resume.png]AnimatedPhysicalModel
+        └─▼[P]PhysicalModel
+          └─▼[N]NotificationListener<LayoutChangedNotification>
+              ▼[I]_InkFeatures-[GlobalKey#00000 ink renderer]
+              └─▼[/icons/inspector/resume.png]AnimatedDefaultTextStyle
+                └─▼[/icons/inspector/textArea.png]DefaultTextStyle
+                  └─▼[/icons/inspector/resume.png]AnimatedBuilder
+                    └─▼[C]CustomMultiChildLayout
+                      ├───▼[L]LayoutId-[<_ScaffoldSlot.body>]
+                      │   └─▼[/icons/inspector/atrule.png]MediaQuery
+                      │     └─▼[C]<bold>Center</bold>
+                      │       └─▶[/icons/inspector/textArea.png]<bold>Text</bold>
+                      ├───▼[L]LayoutId-[<_ScaffoldSlot.appBar>]
+                      │   └─▼[/icons/inspector/atrule.png]MediaQuery
+                      │     └─▼[C]ConstrainedBox
+                      │       └─▼[F]FlexibleSpaceBarSettings
+                      │           ▶[A]<bold>AppBar</bold>
+                      └─▼[L]LayoutId-[<_ScaffoldSlot.floatingActionButton>]
+                        └─▼[/icons/inspector/atrule.png]MediaQuery
+                          └─▼[F]_FloatingActionButtonTransition
+                            └─▼[/icons/inspector/value.png]Stack
+                              └─▼[/icons/inspector/resume.png]ScaleTransition
+                                └─▼[/icons/inspector/colors.png]Transform
+                                  └─▼[/icons/inspector/resume.png]RotationTransition
+                                    └─▼[/icons/inspector/colors.png]Transform
