+++ packages/devtools/test/goldens/inspector_controller_selection_with_styles.txt
@@ -1,8 +1,8 @@
-▼[R]<grayed> </grayed>root <grayed>]</grayed>
-  ▼[M]<grayed> </grayed>MyApp
-    ▼[M]<grayed> </grayed>MaterialApp
-      ▼[S]<grayed> </grayed>Scaffold
-      ├───▼[C]<grayed> </grayed>Center
-      │     [/icons/inspector/textArea.png]<grayed> </grayed>Text <-- selected
-      └─▼[A]<grayed> </grayed>AppBar
-          [/icons/inspector/textArea.png]<grayed> </grayed>Text
+▼[R][root]
+  ▼[M]MyApp
+    ▼[M]MaterialApp
+      ▼[S]Scaffold
+      ├───▼[C]Center
+      │     [/icons/inspector/textArea.png]Text <-- selected
+      └─▼[A]AppBar
+          [/icons/inspector/textArea.png]Text
