+++ packages/devtools/test/goldens/inspector_controller_text_details_tree.txt
@@ -1,14 +1,14 @@
-▼[/icons/inspector/textArea.png] Text <-- selected
-│   "Hello, World!"
-│   textAlign: null [D]
-│   textDirection: null [D]
-│   locale: null [D]
-│   softWrap: null [D]
-│   overflow: null [D]
-│   textScaleFactor: null [D]
-│   maxLines: null [D]
-│   dependencies: [DefaultTextStyle, MediaQuery]
-└─▼[/icons/inspector/textArea.png] RichText
+▼[/icons/inspector/textArea.png]Text <-- selected
+│ "Hello, World!"
+│ textAlign: null [D]
+│ textDirection: null [D]
+│ locale: null [D]
+│ softWrap: null [D]
+│ overflow: null [D]
+│ textScaleFactor: null [D]
+│ maxLines: null [D]
+│ dependencies: [DefaultTextStyle, MediaQuery]
+└─▼[/icons/inspector/textArea.png]RichText
     softWrap: wrapping at box width
     maxLines: unlimited
     text: "Hello, World!"
