+++ packages/devtools/test/goldens/inspector_controller_text_details_tree_with_styles.txt
@@ -1,16 +1,16 @@
-▼[/icons/inspector/textArea.png]<grayed> </grayed><bold>Text</bold> <-- selected
-│   "Hello, World!"
-│   textAlign: null [D]
-│   textDirection: null [D]
-│   locale: null [D]
-│   softWrap: null [D]
-│   overflow: null [D]
-│   textScaleFactor: null [D]
-│   maxLines: null [D]
-│   dependencies: [DefaultTextStyle, MediaQuery]
-└─▼[/icons/inspector/textArea.png]<grayed> </grayed>RichText
+▼[/icons/inspector/textArea.png]<bold>Text</bold> <-- selected
+│ "Hello, World!"
+│ textAlign: null [D]
+│ textDirection: null [D]
+│ locale: null [D]
+│ softWrap: null [D]
+│ overflow: null [D]
+│ textScaleFactor: null [D]
+│ maxLines: null [D]
+│ dependencies: [DefaultTextStyle, MediaQuery]
+└─▼[/icons/inspector/textArea.png]RichText
     softWrap: wrapping at box width
     maxLines: unlimited
     text: "Hello, World!"
     dependencies: [_LocalizationsScope-[GlobalKey#00000], Directionality]
-    ▶renderObject: RenderParagraph#00000 relayoutBoundary=up2
+    ▶renderObject: RenderParagraph<grayed>#00000 relayoutBoundary=up2</grayed>
