+++ packages/devtools/test/goldens/inspector_service_text_details_tree.txt
@@ -13,5 +13,6 @@ Text
      softWrap: wrapping at box width
      maxLines: unlimited
      text: "Hello, World!"
-     dependencies: [_LocalizationsScope-[GlobalKey#00000], Directionality]
+     dependencies: [_LocalizationsScope-[GlobalKey#00000],
+       Directionality]
      renderObject: RenderParagraph#00000 relayoutBoundary=up2
