+++ packages/devtools/test/goldens_stable/inspector_controller_details_tree_scaffold_expanded.txt
@@ -1,29 +1,29 @@
-▼[S] Scaffold <-- selected
-└─▼[S] _ScaffoldScope
-    ▼[P] PrimaryScrollController
-    └─▼[M] Material
-      └─▼[/icons/inspector/resume.png] AnimatedPhysicalModel
-        └─▼[P] PhysicalModel
-          └─▼[N] NotificationListener <LayoutChangedNotification>
-              ▼[I] _InkFeatures [GlobalKey#00000 ink renderer]
-              └─▼[/icons/inspector/resume.png] AnimatedDefaultTextStyle
-                └─▼[/icons/inspector/textArea.png] DefaultTextStyle
-                  └─▼[/icons/inspector/resume.png] AnimatedBuilder
-                    └─▼[C] CustomMultiChildLayout
-                      ├───▼[L] LayoutId [<_ScaffoldSlot.body>]
-                      │   └─▼[/icons/inspector/atrule.png] MediaQuery
-                      │     └─▼[C] Center
-                      │       └─▶[/icons/inspector/textArea.png] Text
-                      ├───▼[L] LayoutId [<_ScaffoldSlot.appBar>]
-                      │   └─▼[/icons/inspector/atrule.png] MediaQuery
-                      │     └─▼[C] ConstrainedBox
-                      │       └─▼[F] FlexibleSpaceBarSettings
-                      │           ▶[A] AppBar
-                      └─▼[L] LayoutId [<_ScaffoldSlot.floatingActionButton>]
-                        └─▼[/icons/inspector/atrule.png] MediaQuery
-                          └─▼[F] _FloatingActionButtonTransition
-                            └─▼[/icons/inspector/value.png] Stack
-                              └─▼[/icons/inspector/resume.png] ScaleTransition
-                                └─▼[/icons/inspector/colors.png] Transform
-                                  └─▼[/icons/inspector/resume.png] RotationTransition
-                                    └─▼[/icons/inspector/colors.png] Transform
+▼[S]Scaffold <-- selected
+└─▼[S]_ScaffoldScope
+    ▼[P]PrimaryScrollController
+    └─▼[M]Material
+      └─▼[/icons/inspector/resume.png]AnimatedPhysicalModel
+        └─▼[P]PhysicalModel
+          └─▼[N]NotificationListener<LayoutChangedNotification>
+              ▼[I]_InkFeatures-[GlobalKey#00000 ink renderer]
+              └─▼[/icons/inspector/resume.png]AnimatedDefaultTextStyle
+                └─▼[/icons/inspector/textArea.png]DefaultTextStyle
+                  └─▼[/icons/inspector/resume.png]AnimatedBuilder
+                    └─▼[C]CustomMultiChildLayout
+                      ├───▼[L]LayoutId-[<_ScaffoldSlot.body>]
+                      │   └─▼[/icons/inspector/atrule.png]MediaQuery
+                      │     └─▼[C]Center
+                      │       └─▶[/icons/inspector/textArea.png]Text
+                      ├───▼[L]LayoutId-[<_ScaffoldSlot.appBar>]
+                      │   └─▼[/icons/inspector/atrule.png]MediaQuery
+                      │     └─▼[C]ConstrainedBox
+                      │       └─▼[F]FlexibleSpaceBarSettings
+                      │           ▶[A]AppBar
+                      └─▼[L]LayoutId-[<_ScaffoldSlot.floatingActionButton>]
+                        └─▼[/icons/inspector/atrule.png]MediaQuery
+                          └─▼[F]_FloatingActionButtonTransition
+                            └─▼[/icons/inspector/value.png]Stack
+                              └─▼[/icons/inspector/resume.png]ScaleTransition
+                                └─▼[/icons/inspector/colors.png]Transform
+                                  └─▼[/icons/inspector/resume.png]RotationTransition
+                                    └─▼[/icons/inspector/colors.png]Transform
