+++ packages/devtools/test/goldens_stable/inspector_controller_initial_tree_with_styles.txt
@@ -1,8 +1,8 @@
-▼[R]<grayed> </grayed>root <grayed>]</grayed>
-  ▼[M]<grayed> </grayed>MyApp
-    ▼[M]<grayed> </grayed>MaterialApp
-      ▼[S]<grayed> </grayed>Scaffold
-      ├───▼[C]<grayed> </grayed>Center
-      │     [/icons/inspector/textArea.png]<grayed> </grayed>Text
-      └─▼[A]<grayed> </grayed>AppBar
-          [/icons/inspector/textArea.png]<grayed> </grayed>Text
+▼[R][root]
+  ▼[M]MyApp
+    ▼[M]MaterialApp
+      ▼[S]Scaffold
+      ├───▼[C]Center
+      │     [/icons/inspector/textArea.png]Text
+      └─▼[A]AppBar
+          [/icons/inspector/textArea.png]Text
