+++ packages/devtools/test/goldens_stable/logging_controller_navigation.txt
@@ -1,0 +1,21 @@
+{
+  "type": "Event",
+  "kind": "Extension",
+  "extensionKind": "Flutter.Navigation",
+  "isolate": {
+    "type": "@Isolate",
+    "id": "isolates/<NUMBER>",
+    "name": "main",
+    "number": "<NUMBER>"
+  },
+  "timestamp": <NUMBER>,
+  "extensionData": {
+    "route": {
+      "description": "MaterialPageRoute<dynamic>(/)",
+      "settings": {
+        "name": "/",
+        "isInitialRoute": true
+      }
+    }
+  }
+}
