+++ packages/devtools/test/goldens_stable/logging_controller_overflow_error.txt
@@ -1,0 +1,41 @@
+Exception Caught By rendering library
+The following assertion was thrown during layout:
+A RenderFlex overflowed by <NUMBER> pixels on the bottom.
+
+The overflowing RenderFlex has an orientation of Axis.vertical.
+The edge of the RenderFlex that is overflowing has been marked in the rendering with a yellow and black striped pattern. This is usually caused by the contents being too big for the RenderFlex.
+Consider applying a flex factor (e.g. using an Expanded widget) to force the children of the RenderFlex to fit within the available space instead of being sized to their natural size.
+This is considered an error condition because it indicates that there is content that cannot be seen. If the content is legitimately bigger than the available space, consider clipping it with a ClipRect widget before putting it in the flex, or using a scrollable container rather than a Flex, like a ListView.
+▼The specific RenderFlex in question is: RenderFlex#<NUMBER> OVERFLOWING
+│ needs compositing
+│ parentData: <none> (can use size)
+│ constraints: BoxConstraints(w=<NUMBER>, h=<NUMBER>)
+│ size: Size(<NUMBER>, <NUMBER>)
+│ direction: vertical
+│ mainAxisAlignment: center
+│ mainAxisSize: max
+│ crossAxisAlignment: center
+│ verticalDirection: down
+├─▼child <NUMBER>: RenderErrorBox#<NUMBER>
+│   parentData: offset=Offset(<NUMBER>, <NUMBER>); flex=null; fit=null (can use size)
+│   constraints: BoxConstraints(<NUMBER><=w<=<NUMBER>, <NUMBER><=h<=Infinity)
+│   size: Size(<NUMBER>, <NUMBER>)
+└─▼child <NUMBER>: RenderSemanticsAnnotations#<NUMBER> relayoutBoundary=up<NUMBER>
+  │   needs compositing
+  │   parentData: offset=Offset(<NUMBER>, <NUMBER>); flex=null; fit=null (can use size)
+  │   constraints: BoxConstraints(<NUMBER><=w<=<NUMBER>, <NUMBER><=h<=Infinity)
+  │   semantic boundary
+  │   size: Size(<NUMBER>, <NUMBER>)
+  └─▼_RenderInputPadding#<NUMBER> relayoutBoundary=up<NUMBER>
+    │   needs compositing
+    │   parentData: <none> (can use size)
+    │   constraints: BoxConstraints(<NUMBER><=w<=<NUMBER>, <NUMBER><=h<=Infinity)
+    │   size: Size(<NUMBER>, <NUMBER>)
+    └─▼RenderConstrainedBox#<NUMBER> relayoutBoundary=up<NUMBER>
+      │   needs compositing
+      │   parentData: offset=Offset(<NUMBER>, <NUMBER>) (can use size)
+      │   constraints: BoxConstraints(<NUMBER><=w<=<NUMBER>, <NUMBER><=h<=Infinity)
+      │   size: Size(<NUMBER>, <NUMBER>)
+      │   additionalConstraints: BoxConstraints(<NUMBER><=w<=Infinity, <NUMBER><=h<=Infinity)
+      └─▶RenderPhysicalShape#<NUMBER> relayoutBoundary=up<NUMBER>
+◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤◢◤
